Set up shared memory between TA and Attestation
We use the existing Camkes primitives of dataports and SharedData
connectors to implement the shared memory between the Trusted
Application component and the Attestation component. The buffer.h header
file will be altered by a separate resizing script dynamically (based on
the size of the ELF binary the TA is compiled into. A remapping script
will map the code pages corresponding to the TA  into the virtual memory
space of the Attestation component. Using the dataport/SharedData
connection allows us to easily address this region in memory. Under the
hood, Camkes sets up the symbols to enable this.