Return signature as null terminated base64 string
In earlier versions, we returned a char pointer that would indeed start
at the memory address range containing the signature, but there was no
good way to determine its size. There might be null bytes in the memory
range too, which complicates interpretation as a string.

Thus instead, we now encode the signature value in base64 and return
this as a properly null-terminated string up through the TEE, until the
Runner receives it and may handle it however it wants.